default_xgboost_container/requirements.txt
pandas==1.5.3
matplotlib==3.7.1
numpy==1.24.4
scikit-learn==1.3
xgboost==1.7.6
joblib==1.3.2
google-cloud-storage>=1.26.0,<2.0.0dev
